+++ app/src/main/java/com/framgia/fdms/data/source/api/service/FDMSApi.java
@@ -143,4 +143,11 @@ public interface FDMSApi {
     @PATCH("api/vendors/{vendor_id}")
     Observable<Respone<List<String>>> updateVendor(@Path("vendor_id") int vendorId,
         @Query("vendor[name]") String name, @Query("vendor[description]") String description);
+
+    @GET("api/device_groups")
+    Observable<Respone<List<Status>>> getDeviceGroups();
+
+    @GET("api/device_categories")
+    Observable<Respone<List<Status>>> getCategoriesByDeviceGroupId(
+        @Query("device_group_id") int deviceGroupId);
 }
